requirements.txt
# Python Dependencies

# Backend Framework
fastapi
python-multipart # For form data parsing in FastAPI
uvicorn[standard] # For running FastAPI server

# Database Connectors
pyodbc # For SQL Server
mysql-connector-python # For MySQL

# Data Handling
pandas

# Machine Learning
sentence-transformers
scikit-learn
# hdbscan # Uncomment if HDBSCAN is chosen

# Environment Variable Management
python-dotenv

# Security (Authentication & Authorization)
passlib[bcrypt] # For password hashing
python-jose[cryptography] # For JWT tokens
email-validator # For Pydantic email validation

# Optional: For scheduling if not using OS-level cron/task scheduler
# apscheduler
